bug fix delete_old

+++ infretis/classes/repex_staple.py
@@ -1153,35 +1153,40 @@ class REPEX_state_staple(REPEX_state):
                 ):
                     if len(self.pn_olds) > self.n - 2:
                         pn_old_del, del_dic = next(iter(self.pn_olds.items()))
-                        load_dir = self.config["simulation"]["load_dir"]
-                        if self.config["output"]["keep_maxop_trajs"]:
-                            path_dir = os.path.join(load_dir, pn_old_del)
-                            # delete trajectory files if low orderp (infinit)
-                            # and directory is not a symlink
-                            if del_dic["max_op"][
-                                0
-                            ] < self.maxop and not os.path.islink(path_dir):
-                                # update maxop and then delete|
+                        # STAPLE-specific safety guard: a path that was
+                        # displaced from one ensemble may still be the active
+                        # trajectory in another (due to infinite swaps).
+                        # Only delete once the path number is no longer live.
+                        if int(pn_old_del) not in set(self.live_paths()):
+                            load_dir = self.config["simulation"]["load_dir"]
+                            if self.config["output"]["keep_maxop_trajs"]:
+                                path_dir = os.path.join(load_dir, pn_old_del)
+                                # delete trajectory files if low orderp (infinit)
+                                # and directory is not a symlink
+                                if del_dic["max_op"][
+                                    0
+                                ] < self.maxop and not os.path.islink(path_dir):
+                                    # update maxop and then delete|
+                                    for adress in del_dic["adress"]:
+                                        os.remove(adress)
+                            else:
+                                # delete trajectory files
                                 for adress in del_dic["adress"]:
                                     os.remove(adress)
-                        else:
-                            # delete trajectory files
-                            for adress in del_dic["adress"]:
-                                os.remove(adress)
-                        # delete txt files
-                        if self.config["output"].get("delete_old_all", False):
-                            for txt in ("order.txt", "traj.txt", "energy.txt"):
-                                txt_adress = os.path.join(
-                                    load_dir, pn_old_del, txt
+                            # delete txt files
+                            if self.config["output"].get("delete_old_all", False):
+                                for txt in ("order.txt", "traj.txt", "energy.txt"):
+                                    txt_adress = os.path.join(
+                                        load_dir, pn_old_del, txt
+                                    )
+                                    if os.path.isfile(txt_adress):
+                                        os.remove(txt_adress)
+                                os.rmdir(
+                                    os.path.join(load_dir, pn_old_del, "accepted")
                                 )
-                                if os.path.isfile(txt_adress):
-                                    os.remove(txt_adress)
-                            os.rmdir(
-                                os.path.join(load_dir, pn_old_del, "accepted")
-                            )
-                            os.rmdir(os.path.join(load_dir, pn_old_del))
-                        # pop the deleted path.
-                        self.pn_olds.pop(pn_old_del)
+                                os.rmdir(os.path.join(load_dir, pn_old_del))
+                            # pop the deleted path.
+                            self.pn_olds.pop(pn_old_del)
                     # keep delete list:
                     if len(self.pn_olds) <= self.n - 2:
                         self.pn_olds[str(pn_old)] = {
